remove "old" slider
Enact-DCO-1.0-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)
