--- packages/moonstone/NewSlider/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "Slider.js" 
-}
