+++ packages/moonstone/Slider/Slider.js
@@ -4,58 +4,42 @@
  * @module moonstone/Slider
  */
 
-import {forKey, forProp, forward, handle, oneOf, stopImmediate} from '@enact/core/handle';
+import {forKey, forProp, forward, forwardWithPrevent, handle} from '@enact/core/handle';
 import kind from '@enact/core/kind';
-import {memoize} from '@enact/core/util';
-import ilib from '@enact/i18n';
-import NumFmt from '@enact/i18n/ilib/lib/NumFmt';
+import Changeable from '@enact/ui/Changeable';
 import Spottable from '@enact/spotlight/Spottable';
-import Pure from '@enact/ui/internal/Pure';
-import Touchable from '@enact/ui/Touchable';
+import Slottable from '@enact/ui/Slottable';
+import UiSlider from '@enact/ui/Slider';
 import PropTypes from 'prop-types';
+import anyPass from 'ramda/src/anyPass';
 import compose from 'ramda/src/compose';
 import React from 'react';
 
-import InternalSliderDecorator from '../internal/SliderDecorator';
-import {computeProportionProgress} from '../internal/SliderDecorator/util';
 import {ProgressBarTooltip} from '../ProgressBar';
 import Skinnable from '../Skinnable';
 
-import SliderBar from './SliderBar';
+import SliderBehaviorDecorator from './SliderBehaviorDecorator';
+import {
+	handleDecrement,
+	handleIncrement
+} from './utils';
 
 import componentCss from './Slider.less';
 
-const isActive = (ev, props) => props.active || props.activateOnFocus || props.detachedKnob;
-const isIncrement = (ev, props) => forKey((props.orientation === 'vertical') ? 'up' : 'right', ev);
-const isDecrement = (ev, props) => forKey((props.orientation === 'vertical') ? 'down' : 'left', ev);
 
-// memoize percent formatter for each locale so that we only instantiate NumFmt when locale changes
-const memoizedPercentFormatter = memoize((/* locale */) => new NumFmt({type: 'percentage', useNative: false}));
 
-/**
- * A stateless range-selection input. In most circumstances, you will want to use the stateful
- * version: {@link moonstone/Slider.Slider}.
+/* ***************************************************
+ * Still need to reconcile current API
+ * * consider multiple knob and bar support
+ * * integration with IncrementSlider
+ * * integration with MediaSlider
  *
- * @class SliderBase
- * @memberof moonstone/Slider
- * @ui
- * @public
- */
+ * ***************************************************/
+
 const SliderBase = kind({
 	name: 'Slider',
 
-	propTypes: /** @lends moonstone/Slider.SliderBase.prototype */{
-
-		/**
-		 * Overrides the `aria-valuetext` for the slider. By default, `aria-valuetext` is set
-		 * to the current value. This should only be used when the parent controls the value of
-		 * the slider directly through the props.
-		 *
-		 * @type {String|Number}
-		 * @memberof moonstone/Slider.SliderBase.prototype
-		 * @public
-		 */
-		'aria-valuetext': PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+	propTypes: {
 
 		/**
 		 * When `true`, the component may be manipulated via the directional input keys upon
@@ -84,15 +68,6 @@ const SliderBase = kind({
 		backgroundProgress: PropTypes.number,
 
 		/**
-		 * The custom value or component for the tooltip. If [tooltip]{@link moonstone/Slider.SliderBase#tooltip},
-		 * is `true`, then it will use built-in tooltip with given a string. If `false`, a custom tooltip
-		 * component, which follows the knob, may be used instead.
-		 *
-		 * @type {String|Node}
-		 */
-		children: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-
-		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
 		 * corresponding internal Elements and states of this component.
 		 *
@@ -100,16 +75,6 @@ const SliderBase = kind({
 		 * @private
 		 */
 		css: PropTypes.object,
-
-		/**
-		 * The slider can change its behavior to have the knob follow the cursor as it moves
-		 * across the slider, without applying the position. A click or drag behaves the same.
-		 * This is primarily used by media playback. Setting this to `true` enables this behavior.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		detachedKnob: PropTypes.bool,
 
 		/**
 		 * When `true`, the component is shown as disabled and does not generate events
@@ -125,24 +90,20 @@ const SliderBase = kind({
 		 * @public
 		 */
 		focused: PropTypes.bool,
+		knob: PropTypes.node,
 
 		/**
-		 * The method to run when the input mounts, giving a reference to the DOM.
+		 * The amount to increment or decrement the position of the knob via 5-way controls.
+		 * When `detachedKnob` is false, the knob must first be activated by selecting it. When
+		 * `detachedKnob` is true, the knob will respond to direction key presses without
+		 * activation.
 		 *
-		 * @type {Function}
-		 * @private
+		 * If not specified, `step` is used for the default value.
+		 *
+		 * @type {Number}
+		 * @public
 		 */
-		inputRef: PropTypes.func,
-
-		/**
-		* When not `vertical`, determines which side of the knob the tooltip appears on.
-		* When `false`, the tooltip will be on the left side, when `true`, the tooltip will
-		* be on the right.
-		*
-		* @type {String}
-		* @private
-		*/
-		knobAfterMidpoint: PropTypes.bool,
+		knobStep: PropTypes.number,
 
 		/**
 		 * The maximum value of the slider.
@@ -189,55 +150,13 @@ const SliderBase = kind({
 		onChange: PropTypes.func,
 
 		/**
-		 * The handler to run when the value is decremented.
+		 * If `true` the slider will be oriented vertically.
 		 *
-		 * @type {Function}
-		 * @param {Object} event
+		 * @type {Boolean}
+		 * @default false
 		 * @public
 		 */
-		onDecrement: PropTypes.func,
-
-		/**
-		 * The handler to run when the value is incremented.
-		 *
-		 * @type {Function}
-		 * @param {Object} event
-		 * @public
-		 */
-		onIncrement: PropTypes.func,
-
-		/**
-		 * The handler to run when the knob moves. This method is only called when running
-		 * `Slider` with `detachedKnob`. If you need to run a callback without a detached knob
-		 * use the more traditional `onChange` property.
-		 *
-		 * @type {Function}
-		 * @param {Object} event
-		 * @param {Number} event.proportion The proportional position of the knob across the slider
-		 * @param {Boolean} event.detached `true` if the knob is currently detached, `false` otherwise
-		 * @public
-		 */
-		onKnobMove: PropTypes.func,
-
-		/**
-		 * The handler to run when the mouse is moved across the slider.
-		 *
-		 * @type {Function}
-		 * @param {Object} event
-		 * @param {Number} event.value Value of the slider
-		 * @public
-		 */
-		onMouseMove: PropTypes.func,
-
-		/**
-		 * Sets the orientation of the slider, whether the slider moves left and right or up and
-		 * down. Must be either `'horizontal'` or `'vertical'`.
-		 *
-		 * @type {String}
-		 * @default 'horizontal'
-		 * @public
-		 */
-		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
+		orientation: PropTypes.string,
 
 		/**
 		 * When `true`, a pressed visual effect is applied
@@ -246,32 +165,6 @@ const SliderBase = kind({
 		 * @public
 		 */
 		pressed: PropTypes.bool,
-
-		/**
-		 * `scrubbing` only has an effect with a detachedKnob, and is a performance optimization
-		 * to not allow re-assignment of the knob's value (and therefore position) during direct
-		 * user interaction.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		scrubbing: PropTypes.bool,
-
-		/**
-		 * The method to run when the slider bar component mounts, giving a reference to the DOM.
-		 *
-		 * @type {Function}
-		 * @private
-		 */
-		sliderBarRef: PropTypes.func,
-
-		/**
-		 * The method to run when mounted, giving a reference to the DOM.
-		 *
-		 * @type {Function}
-		 * @private
-		 */
-		sliderRef: PropTypes.func,
 
 		/**
 		 * The amount to increment or decrement the value.
@@ -292,39 +185,6 @@ const SliderBase = kind({
 		tooltip: PropTypes.bool,
 
 		/**
-		 * Converts the contents of the built-in tooltip to a percentage of the bar.
-		 * The percentage respects the min and max value props.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		tooltipAsPercent: PropTypes.bool,
-
-		/**
-		 * Setting to `true` overrides the natural LTR->RTL tooltip side-flipping for locale
-		 * changes for `vertical` sliders. This may be useful if you have a static layout that
-		 * does not automatically reverse when in an RTL language.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		tooltipForceSide: PropTypes.bool,
-
-		/**
-		 * Specify where the tooltip should appear in relation to the Slider bar. Options are
-		 * `'before'` and `'after'`. `before` renders above a `horizontal` slider and to the
-		 * left of a `vertical` Slider. `after` renders below a `horizontal` slider and to the
-		 * right of a `vertical` Slider. In the `vertical` case, the rendering position is
-		 * automatically reversed when rendering in an RTL locale. This can be overridden by
-		 * using the [tooltipForceSide]{@link moonstone/Slider.Slider.tooltipForceSide} prop.
-		 *
-		 * @type {String}
-		 * @default 'before'
-		 * @public
-		 */
-		tooltipSide: PropTypes.oneOf(['before', 'after']),
-
-		/**
 		 * The value of the slider.
 		 *
 		 * @type {Number}
@@ -337,28 +197,13 @@ const SliderBase = kind({
 	defaultProps: {
 		activateOnFocus: false,
 		active: false,
-		backgroundProgress: 0,
-		knobAfterMidpoint: false,
-		detachedKnob: false,
-		focused: false,
-		max: 100,
-		min: 0,
-		noFill: false,
-		onChange: () => {}, // needed to ensure the base input element is mutable if no change handler is provided
-		orientation: 'horizontal',
-		pressed: false,
-		step: 1,
-		tooltip: false,
-		tooltipAsPercent: false,
-		tooltipForceSide: false,
-		tooltipSide: 'before',
-		value: 0
+		disabled: false
 	},
 
 	styles: {
 		css: componentCss,
 		className: 'slider',
-		publicClassNames: true
+		publicClassNames: ['slider']
 	},
 
 	handlers: {
@@ -368,165 +213,57 @@ const SliderBase = kind({
 			forward('onActivate')
 		),
 		onKeyDown: handle(
-			forward('onKeyDown'),
-			isActive,
-			oneOf(
-				[isDecrement, forward('onDecrement')],
-				[isIncrement, forward('onIncrement')]
-			),
-			stopImmediate
+			forwardWithPrevent('onKeyDown'),
+			anyPass([
+				handleIncrement,
+				handleDecrement
+			])
 		),
 		onKeyUp: handle(
-			forward('onKeyUp'),
+			forwardWithPrevent('onKeyUp'),
 			forProp('activateOnFocus', false),
 			forKey('enter'),
 			forward('onActivate')
-		),
-		onMouseUp: handle(
-			forward('onMouseUp'),
-			(ev) => {
-				// This bit of hackery allows us to use the <input> for dragging but sends the
-				// focus back to the component when the pointer is released. It works because,
-				// for some reason, the <input> still sends a mouseup event even when the
-				// pointer is released while off the <input>.
-				if (ev.target.nodeName === 'INPUT') {
-					ev.currentTarget.focus();
-				}
-			}
 		)
 	},
 
 	computed: {
-		children: ({children, max, min, tooltip, tooltipAsPercent, value}) => {
-			if (!tooltip || children) return children;
+		className: ({activateOnFocus, active, styler}) => styler.append({
+			activateOnFocus,
+			active
+		}),
+		tooltipComponent: ({focused, tooltip}) => {
+			if (!focused) return null;
+			if (tooltip === true) return ProgressBarTooltip;
 
-			if (tooltipAsPercent) {
-				const formatter = memoizedPercentFormatter(ilib.getLocale());
-				const percent = Math.floor(computeProportionProgress({value, max, min}) * 100);
-
-				return formatter.format(percent);
-			}
-
-			return value;
-		},
-		className: ({activateOnFocus, active, noFill, orientation, pressed, styler}) => styler.append(
-			orientation,
-			{
-				activateOnFocus,
-				active,
-				noFill,
-				pressed
-			}
-		),
-		proportionProgress: computeProportionProgress
+			return tooltip;
+		}
 	},
 
-	render: ({backgroundProgress, children, css, disabled, focused, inputRef, knobAfterMidpoint, max, min, onBlur, onChange, onKeyDown, onMouseMove, onMouseUp, orientation, proportionProgress, scrubbing, sliderBarRef, sliderRef, step, tooltip, tooltipForceSide, tooltipSide, value, ...rest}) => {
-		delete rest.activateOnFocus;
-		delete rest.active;
-		delete rest.detachedKnob;
-		delete rest.noFill;
-		delete rest.onActivate;
-		delete rest.onDecrement;
-		delete rest.onIncrement;
-		delete rest.onKnobMove;
-		delete rest.pressed;
-		delete rest.tooltipAsPercent;
-
-		let tooltipComponent = null;
-
-		// when `tooltip` is `false`, use custom tooltip provided in `children`
-		if (!tooltip) {
-			tooltipComponent = children;
-		} else if (focused) {
-			// only display tooltip when `focused`
-			tooltipComponent = <ProgressBarTooltip
-				knobAfterMidpoint={knobAfterMidpoint}
-				forceSide={tooltipForceSide}
-				orientation={orientation}
-				proportion={proportionProgress}
-				side={tooltipSide}
-			>
-				{children}
-			</ProgressBarTooltip>;
-		}
+	render: (props) => {
+		delete props.activateOnFocus;
+		delete props.active;
+		delete props.focused;
+		delete props.tooltip;
 
 		return (
-			<div
-				{...rest}
-				aria-disabled={disabled}
-				disabled={disabled}
-				onBlur={onBlur}
-				onKeyDown={onKeyDown}
-				onMouseUp={onMouseUp}
-				ref={sliderRef}
-			>
-				<SliderBar
-					css={css}
-					orientation={orientation}
-					proportionBackgroundProgress={backgroundProgress}
-					proportionProgress={proportionProgress}
-					ref={sliderBarRef}
-					scrubbing={scrubbing}
-				>
-					{tooltipComponent}
-				</SliderBar>
-				<input
-					aria-disabled={disabled}
-					className={css.input}
-					disabled={disabled}
-					type="range"
-					ref={inputRef}
-					max={max}
-					min={min}
-					step={step}
-					onChange={onChange}
-					onMouseMove={onMouseMove}
-					orient={orientation}
-					value={value}
-				/>
-			</div>
+			<UiSlider
+				{...props}
+				css={props.css}
+			/>
 		);
 	}
 });
 
 const SliderDecorator = compose(
-	Pure,
+	Changeable,
+	SliderBehaviorDecorator,
 	Spottable,
-	InternalSliderDecorator,
-	Touchable({activeProp: 'pressed'}),
+	Slottable({slots: ['knob', 'tooltip']}),
 	Skinnable
 );
 
-/**
- * Range-selection input with Moonstone styling, Spottable, Touchable and SliderDecorator applied.
- *
- * By default, `Slider` maintains the state of its `value` property. Supply the `defaultValue`
- * property to control its initial value. If you wish to directly control updates to the
- * component, supply a value to `value` at creation time and update it in response to `onChange`
- * events.
- *
- * @class Slider
- * @memberof moonstone/Slider
- * @mixes ui/Touchable.Touchable
- * @mixes spotlight/Spottable.Spottable
- * @ui
- * @public
- */
 const Slider = SliderDecorator(SliderBase);
-
-/**
- * A {@link moonstone/TooltipDecorator.Tooltip} specifically adapted for use with
- * {@link moonstone/IncrementSlider.IncrementSlider}, {@link moonstone/ProgressBar.ProgressBar},
- * or {@link moonstone/Slider.Slider}.
- *
- * See {@link moonstone/ProgressBar.ProgressBarTooltip}
- *
- * @class SliderTooltip
- * @memberof moonstone/Slider
- * @ui
- * @public
- */
 
 export default Slider;
 export {
