+++ packages/moonstone/Slider/SliderBehaviorDecorator.js
