+++ packages/moonstone/Slider/SliderTooltip.js
