+++ packages/moonstone/Slider/SliderTooltip.less
